+++ src/components/page-transition/index.ts
@@ -1,0 +1,1 @@
+export { PageTransition } from './page-transition'
